SECURITY FIX: Remove test.php file
test,php reveals a lot of sensible server info by just calling

    <?php phpinfo(); ?>

and can simply get called from anywhere by anyone going to the URL:

    https://openprinting.org/test.php

This is probably a stray debugging file which got included
accidentally. This commit removes it.

Thanks, Ali Adam, for investigating the security of the OpenPrinting
web server and reporting this bug.